feat: Alter PK columns type to IDENTITY (#60)
* feat: Updated accounts and users type to identity

* fix: Failing code

* feat: Updated items table

* feat: Updated user_orders table

* feat: Updated receipts table

* fix: Receipt schema change

* fix: Failing tests

+++ .gen/public/model/items.go
@@ -12,7 +12,7 @@ import (
 )
 
 type Items struct {
-	ID        int32 `sql:"primary_key"`
+	ID        int64 `sql:"primary_key"`
 	ReceiptID int64
 	Price     int32
 	Name      *string
